RSDEV-909: move dataverse bean out of prod profile so that devs may use Dataverse
